--- lib/models/company_manager.dart
@@ -1,18 +1,0 @@
-// import 'package:flutter/material.dart';
-// import 'package:firebase_core/firebase_core.dart';
-// import 'package:elli_admin/firebase_handler.dart';
-// import 'package:elli_admin/theme_elicit.dart';
-//
-// class CompanyManager extends StatefulWidget {
-//   const CompanyManager({Key? key}) : super(key: key);
-//
-//   @override
-//   State<StatefulWidget> createState() => _CompanyManagerState();
-// }
-//
-// class _CompanyManagerState extends State<CompanyManager> {
-//   @override
-//   Widget build(BuildContext context) {
-//     return Text("hej");
-//   }
-// }
